requirements.txt
appnope==0.1.4  # macOS specific, consider removing if not needed
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.1
cryptography==42.0.2
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
fonttools==4.48.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
numpy==1.26.4
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip-tools==7.4.0
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
# pywin32==306  # Removed, Windows-specific
pyzmq==25.1.2
# referencing==0.33.0  # This package does not seem common, ensure it's needed
requests==2.31.0
# rpds-py==0.18.0  # Not found in PyPI, verify if this is correct
setuptools==69.1.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
yarg==0.1.9
